+++ pyssata/lib/layers2pupil_ef.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pyssata import xp
-from pyssata import standard_dtype
+from pyssata import float_dtype
 
 from scipy.ndimage import rotate
 import warnings
@@ -19,7 +19,7 @@ def single_layer2pupil_ef(layer_ef, polar_coordinate, height_source, update_ef=N
     diff_height = height_source - height_layer
 
     if (height_layer == 0 or (xp.isinf(height_source) and polar_coordinate[0] == 0)) and \
-       ((shiftXY is None) or (xp.all(shiftXY == xp.array([0, 0], dtype=standard_dtype)))) and \
+       ((shiftXY is None) or (xp.all(shiftXY == xp.array([0, 0], dtype=float_dtype)))) and \
        ((pupil_position is None) or (pupil_position == 0)) and \
        ((rotAnglePhInDeg is None) or (rotAnglePhInDeg == 0)) and \
        ((magnify is None) or (magnify == 1)):
