+++ pyssata/processing_objects/int_control.py
@@ -4,12 +4,12 @@ from pyssata.lib.int2iirfilter import int2iirfilter
 import numpy as np
 
 from pyssata import xp
-from pyssata import standard_dtype
+from pyssata import float_dtype
     
 class IntControl(IIRControl):
     def __init__(self, int_gain, ff=None, delay=0, offset=None, og_shaper=None):
         
-        iirfilter = int2iirfilter(xp.array(int_gain, dtype=standard_dtype), ff=ff)
+        iirfilter = int2iirfilter(xp.array(int_gain, dtype=float_dtype), ff=ff)
 
         # Initialize IIRControl object
         super().__init__(iirfilter, delay=delay)
